Add test skeleton for NPLinkedList
Added a basic skeleton for unit testing NPLinkedList.  Also added the same skeleton
to the implementation file.

UnitTests are now setup in the project and follow the same directory conventions as the project.

The test naming convention is as follows
- (void)test_{m,p}_{returntype}_{selector}[_...]

m = minus (instance method)
p = plus  (class method)
returntype = The return type without the pointer character
selector = the selector

Each ":" is replaced with a "_" and the arguments are omitted.

None of the methods or tests are implemented yet.