Add merge sort to NPLinkedList
Added the ability to sort a linked list.  It uses the merge sort
algorithm.  I've added an extra method that performs the merging using a
comparator.  This isn't exposed in the header file as I didn't feel it
would be a useful function to have outside the realms of a merge sort.

The class method that was meant to sort a linked list destructively was
removed due to the nature of the recursive algorithm and the necessity
for secondary temporary lists.  This can probably be added later using
another algorithm.

There are two sorting methods, both return a new linked list.  One is
just a convenience method that calls the other.  One is a class method
the other operates on the calling object.  Neither of them are
destructive.